intro background
