+++ config/config.ini
@@ -26,3 +26,4 @@ down = short the corn!,goblin town,bearish?,dooom,sell!!
 
 [first_run]
 enabled=true
+intro_background_image=bitbot.png
